ci: Update Fedora CI image to Fedora 41
Fedora 39 is out of support so, as per our policy, update the CI image
to the oldest still-supported release, which is 41.

systemtap-sdt-dtrace needs to be installed separately (for the `dtrace`
binary), as it appears to have been split out of some other package.

Update the mingw CI image too, as it’s built on top of the Fedora one.

Update the supported platforms documentation.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>