introspection: Remove backward compatibility from GLib-2.0
As commit db96cb45c:

The identifiers and types in GLib-2.0 that are used for backward
compatibility purposes in order to keep code importing GLib and using
platform-specific API are actually breaking the platform-specific
introspection data, because the introspection scanner favours types
found inside dependencies as opposed to types with the same name found
inside the current namespace.

In practice, the backward compatibility hack of keeping duplicate types
inside GLib-2.0 is effectively preventing people from using GLibUnix-2.0
and GLibWin32-2.0.

We cannot change the introspection scanner, because that could have
unforeseen results across the stack.

In practice, the only reasonable way forward is to break backward
compatibility, and remove the duplicate symbols and identifiers from
GLib-2.0, as we should have done in the first place.