gregex: Handle PCRE2_ERROR_MISSING_NUMBER_TERMINATOR if defined
When the error is defined in a new enough PCRE2 we should handle it.

However let's not define an error message for this in this commit (just
let's use the old one that is generic enough), so that it can be backported
to stable versions without having to require new translations.

Closes: #3809