ci: Track changes to gir files using artifacts and MRs comments
Introspection files are often updated as per the changes we do in
code, but it's really hard to track what gets changed at times, and so
regressions are likely to happen (as gjs#711 or the whole #3744
story).

So, while the files are quite big for being tracked inside the
repository we can still track them using the gitlab artifacts and using
comments of the MRs to expose the diff to the reviewers.

Sadly this is a bit more complex than expected [1], given that gitlab
does not allow to export (for a good reason) any token to pipelines
running in non-protected branches, to avoid leaking them.

So, to still be allow to do this safely we need to use a bit more
complex setup:
 - A new project access token has been generated and exported as
   $GIR_CHECK_TOKEN
 - A CI/CD pipeline trigger token has been created in settings
 - A new webhook has been configured to run on merge request events and
   that triggers the pipeline on main branch using the above pipeline
   trigger token and exposing various events as CI_GLIB_* variables

Now, when a new merge request event happens:
 - A pipeline is triggered if the `CI_GLIB_*` env variables are matching
   the expected ones
 - In such pipeline, the check-introspection jobs are started
 - They look for the job that is matching the said MR
 - We wait for the job being completed and we download the artifacts it
   produces
 - The gir files in the artifacts are compared with the ones in the
   latest artifact we have for the main branch
 - If a diff is found, comments are created (or updated) inside the MR
   containing the diff (in a detail spoiler).

All this required some more advanced playing with the gitlab APIs, and
so that's all done in the alpine image, via python-gitlab.

[1] https://gitlab.gnome.org/GNOME/glib/-/merge_requests/4880#note_2583072