gunixmounts: fix out of bounds write in AIX mount table parser
A mount table is inherently trusted, so this isn't really a security
issue, but let's not write out of bounds. If anybody is still using GLib
on AIX, then you're welcome.

Discovery and patch credit: mhzcyber - Mohammad Hussam Alzeyyat

(I, Michael, only wrote this commit message and adjusted the
whitespace.)