gio/gunixmounts: use more efficient string set detection
This speeds up the parsing of mount entries by a bit more tha 20%.

This is possible by avoiding string set detection using arrays of
strings and instead using a single long string with embedded \0.
Simply check the input string and its trailing \0 in the haystack.

This turns out to not only be faster than doing all of the strcmp()
(which requires looking at the search string repeatedly) but also
faster than using gperf to generate a perfect hash. It also doesn't
require a gperf dependency.

This can be useful on systems with many thousands of mounts.