ci: Enable the gobject-introspection build for the debian-stable-i386 job
So the introspection tests which check architecture-specific things are
run on an i386 architecture. This gives the best check we have that the
libgirepository code is actually architecture independent.

This requires adding some missing dependencies to the Debian Dockerfile
which gobject-introspection.git needs. It also means enabling `sudo` on
those containers.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>