ci: Clean up gobject-introspection build artifacts
I don’t see there’s any reason for us to keep and expose artifacts from
building gobject-introspection as a dependency. The introspection files
subsequently generated for GLib are/can be exposed separately as
artifacts.

In particular, keeping the gobject-introspection directory around as a
subdirectory of the glib.git checkout means that `reuse.sh` starts
investigating the licenses of the gobject-introspection.git files and
failing as a result.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>