gbytes: Fix integer overflows in precondition checks
There were a couple of integer overflows in the precondition checks for
`g_bytes_new_from_bytes()` which could lead to a corrupt `GBytes` being
constructed if the caller passed invalid arguments. Specifically, a
`length` which is big enough to wrap around 2^64-1 when added to
`bytes`.

Fix the checks so they correctly abort the process on such invalid
arguments.

The caller is (and was always) responsible for checking the arguments
before passing them to `g_bytes_new_from_bytes()`. This is not a
security bug. The change in this commit merely makes GLib abort with a
critical warning when previously it didn’t.

Spotted by Codean Labs.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Fixes: #3843