gatomic/MSVC: Use volatile for plain C loads / stores
volatile guarantees that load / store tearing won't happen and that
actual atomic opcodes are used as long as the underlying ISA permits
that. Generally, whenever sizeof (operand) is less than or equal to
sizeof (void *). GCC also extends that to 64 bit load / stores on
x86, but MSVC doesn't.

This is a compiler extension. The C / C++ standard does not mandate
atomicity of volatile loads / stores.