docs(GResource): Tweak overlays description
The documentation about G_RESOURCE_OVERLAYS said
that the "path after the slash" must ideally be absolute, but
a path after a slash is always absolute. The intent seems to
have been that "the path after the `=`" (IOW the filesystem
path being used as a substitution) should be, but is not
required to be, absolute.