inotify: Explicitly handle empty queue in dispatch
This was highlighted by a `-Wsign-conversion` warning.

The return value from `ik_source_get_dispatch_time()` is signed, and can
be `-1`, but the code in `ik_source_dispatch()` was implicitly
converting it to unsigned. This was converting `-1` to a huge positive
integer. This would then lose against `boredom_time` in the `MIN()`
call.

Because of this, the previous behaviour was correct, but potentially
only by accident. Fix the signedness of the types (fixing the
`-Wsign-conversion` warning) and explicitly handle the case of
`ik_source_get_dispatch_time()` returning `-1` correctly, to make the
expected behaviour more obvious.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #3405