gicon: Add an upper bound on the number of tokens in an icon
Highlighted by a `-Wsign-conversion` warning, it was possible for a
tokenised icon to contain enough tokens to overflow the logic in
`g_icon_new_from_tokens()`.

Add an upper bound and return an error if that’s hit, to avoid this.

In practice this can only be hit by calling `g_icon_new_for_string()` on
a string which starts with the correct magic, followed by 2GB of spaces.
This isn’t considered exploitable as it’s unlikely that
`g_icon_new_for_string()` will be called on untrusted input, or that the
untrusted input would not have its length validated beforehand at all.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #3405