gemblem: Validate origin when parsing a GEmblem icon from a string
Spotted by enabling `-Wsign-conversion`. This fixes the
`-Wsign-conversion` warning.

The detokenisation code for `GEmblem` was not validating the origin enum
when parsing it from the tokens, so could happily be passed garbage.
This was unlikely to cause any further problems as the origin is just
passed around in the code verbatim. It would just confuse application
code.

Validate the origin and return an error if it’s invalid. Add a basic
unit test for this parsing.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #3405