gbufferedinputstream: Improve the documentation for peek()
It wasn’t particularly helpful, giving the types of the arguments rather
than their function, and using confusing wording in the docs body which
made it unclear whether `offset` applied to the buffered input or to the
`buffer` argument.

It also said `-1` could be returned on error, despite the return type
not supporting that and none of the code doing that. Drop that clause.

Port to gi-docgen format at the same time.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>