gdesktopappinfo: Error if mimeapps.list is too long
This is unfortunately necessary because `g_key_file_to_data()` can
potentially return a string as long as `G_MAXSIZE`, but
`g_file_set_contents_full()` takes a signed `gssize` as its length
argument.

We can’t change either of those things, and by enabling
`-Wsign-conversion` we can no longer ignore the potential bug, so let’s
throw a runtime error in the very unlikely case that this is hit.

I can’t add a unit test for it because generating a `mimeapps.list`
which is that long would be ludicrous.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #3405