gicon: Guarantee num_tokens is non-negative
It was perhaps an oversight when defining this API to make it an `int`
rather than a `size_t`. The value of `num_tokens` is never negative.
Document that, and add assertions in all the implementations of the
`GIcon` interface to make that clearer.

Use those assertions to fix `-Wsign-conversion` warnings in the
interface implementations where possible.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #3405