gio: Don’t use g_assert_*() in runtime code
Use `g_assert()` instead. This is the opposite of the changes I’ve been
making to the unit tests for a long time to make them use `g_assert_*()`
rather than `g_assert()`.

`g_assert()` will be compiled out of runtime code with
`G_DISABLE_ASSERT`, which is what we want. `g_assert_*()` use the
`GTest` logging functions on failure, which should not be used by
runtime code as they will probably end up sending the messages through
something other than the global `GLogWriterFunc`.

Some of these instances should potentially be returning a `GError` to
the calling code at runtime, but that can be reassessed in future work.
For the moment, I’m just replacing one assertion with another (or a
`g_error()`) which is essentially like-for-like.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>