build: Add -Wshorten-64-to-32 warning arguments
As with `-Wsign-conversion`, these will be enabled in sections across
the codebase, as there will be a lot of warnings to fix.

To begin with, enable `-Wshorten-64-to-32` for the gthread subdirectory.
As it happens, that doesn’t trigger any new warnings. Happy days.

`-Wshorten-64-to-32` warns if a 64-bit number is assigned to a 32-bit
wide storage location. It’s only available on clang.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #3527