gclosure: Fix 64-vs-32 bit integer storage sizes
Fix an array iteration to correctly use `size_t`, and add an assertion
to check the number of arguments passed to libffi is not too many to
overflow its argument (which is, unfortunately, an `unsigned int` rather
than a `size_t`).

Neither of these are real bugs, but fixing the warnings means we can
enable the warning flag permanently to catch actual bugs in future.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #3527