genums: Add explicit casts for enum/flags marshalling with GValue
For historic reasons (probably to try and support enums with more than
32 values), enums and flags are stored in the `long` field of a `GValue`.

This is fine, but it triggers a load of `-Wshorten-64-to-32` warnings.
Add some explicit casts to silence those warnings. I don’t think we can
do any better for the implementation anyway. This doesn’t change the
behaviour from before, it just makes the implicit casts explicit.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #3527