gobject: Add restriction on number of construct properties
The `GObjectClass.n_construct_properties` member is of type `size_t`,
which is correct. However, it’s passed to `GObjectClass.constructor()`,
which takes an `unsigned int`. So in practice its maximum value is
`UINT_MAX`.

Add documentation and an assertion to check that, since it’s public API
and can’t be changed.

In practice, we don’t expect any code to come anywhere near this, but
it’s good to be correct (and this silences a `-Wshorten-64-to-32`
warning).

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #3527