gconvert: Avoid a pointless signed/unsigned switch
Rather than take a `gssize` and allow `-1` to indicate “call `strlen()`
yourself” on this internal function, let’s take a `size_t` and require
the caller to call `strlen()` themselves.

This avoids converting between `size_t` and `gssize`, all the
signed/unsigned comparison warnings that entails, and dropping half of
the available value domain in the process.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>