gbsearcharray: Fix a -Wshorten-64-to-32 warning
The `distance` is calculated as a `size_t` (as it should be, as a memory
offset), but the array only handles indices as `unsigned int`s, so we
need to check and do an explicit shortening cast.

This shouldn’t cause any behavioural differences, as the array would
have to hold millions of items to hit this. Better safe than sorry,
though.

While I’m not working on `-Wshorten-64-to-32` in `glib/` yet, this
header is used internally in other parts of glib.git, so it’s good to
get it warning-free early.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #3527